Moved the solution methods to a different file, renamed the_show to solve_puzzle
